+++ lib/index.ts
@@ -2,3 +2,5 @@ export * from './redis.module';
 export * from './redis.decorators';
 export * from './redis.interfaces';
 export * from './redis.utils';
+export * from './indicator/redis-health.indicator';
+export * from './indicator/redis-health.module';