+++ lib/indicator/redis-health.indicator.ts
@@ -1,0 +1,21 @@
+import { Injectable, Inject } from '@nestjs/common';
+import { HealthIndicator, HealthIndicatorResult, HealthCheckError } from '@nestjs/terminus';
+import Redis from 'ioredis';
+
+import { REDIS_HEALTH_INDICATOR } from '../redis.constants';
+
+@Injectable()
+export class RedisHealthIndicator extends HealthIndicator {
+  constructor(@Inject(REDIS_HEALTH_INDICATOR) private readonly redis: Redis) {
+    super();
+  }
+
+  async isHealthy(key: string): Promise<HealthIndicatorResult> {
+    try {
+      await this.redis.ping();
+      return this.getStatus(key, true);
+    } catch (error) {
+      throw new HealthCheckError('Redis check failed', this.getStatus(key, false, { message: error.message }));
+    }
+  }
+}