+++ lib/redis.constants.ts
@@ -1,3 +1,5 @@
 export const REDIS_MODULE_CONNECTION = 'default';
 export const REDIS_MODULE_CONNECTION_TOKEN = 'IORedisModuleConnectionToken';
 export const REDIS_MODULE_OPTIONS_TOKEN = 'IORedisModuleOptionsToken';
+
+export const REDIS_HEALTH_INDICATOR = 'redisHealthIndicator';