+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.2.3](https://github.com/nest-modules/ioredis/compare/v1.2.2...v1.2.3) (2023-12-21)
+
+### [1.2.2](https://github.com/nest-modules/ioredis/compare/v1.2.1...v1.2.2) (2023-12-21)
+
+
+### Bug Fixes
+
+* correct the use of the redis health indicator ([8f61004](https://github.com/nest-modules/ioredis/commit/8f61004f2984f5ae2949b6664efcc004071fced7))
+
+### [1.2.1](https://github.com/nest-modules/ioredis/compare/v1.2.0...v1.2.1) (2023-12-21)
+
+## [1.2.0](https://github.com/nest-modules/ioredis/compare/v1.1.2...v1.2.0) (2023-12-21)
+
+
+### Features
+
+* add redis health indicator ([ffc2ad7](https://github.com/nest-modules/ioredis/commit/ffc2ad7c49cd1b68a519db9d3bf81c582e562b2d))
+
 ### [1.1.2](https://github.com/nest-modules/ioredis/compare/v1.1.1...v1.1.2) (2023-12-21)
 
 ### [1.1.1](https://github.com/nest-modules/ioredis/compare/v1.1.0...v1.1.1) (2023-12-20)